+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 
 
+## 0.2.0 (February 17, 2017)
+
+* add type from inputSchema to project 
+* automatic generate schema
+* automatic generate models
+* automatic generate resolvers
+
 ## 0.1.0 (February 16, 2017)
 
 * init graphql server project from template
+
+